+++ CHANGELOG.rst
@@ -10,6 +10,19 @@ only be introduced in major versions with advance notice in the **Deprecated**
 section of releases.
 
 
+v3.15.0 (2021-??-??)
+--------------------
+
+Added
+^^^^^
+- Decorator for final classes and an is_final_class function to test it.
+- Support for final classes as type hint.
+
+Fixed
+^^^^^
+- Parsing of relative paths inside inner configs for type hint actions.
+
+
 v3.14.0 (2021-06-08)
 --------------------